Bugfix: Don't reopen an already closed socket
Sockets should be opened explicitly. There was a bug in the case the
peer has reset the connection, and in the local error handling an alert
was setup to be sent, which in the socket module led to the reopening of
the socket (and then even without the endpoint argument).